Moved all the depreciated functs into a new class and file Stored this file on the server side dir: file: generate_summary.py class: Create_Summary
In doing this we can have a distinct seperation between the def's that are required to
generate the cve_summary dict in docker-scanner (at scantime - docker-scanner.py line 450) and the post analysis def's that
are used in xmlp.  Since we now have the new docker_state_file that holds all needed information
the previous def's are no longer used by post-scaning analysis

Creating this seperation is going to cause a need for the sample_api to be re-written with the
newer xmlp file to utilize the pprint and a smaller summarize def (will need to be created)
